sdm710-common: init.qcom.usb.rc: Fix fallback to default composition from RNDIS
If the sys.usb.state is set to any other value other than
rndis(rndis,adb) when tethering is enabled through UI, the
USB device manager falls back to the default or previous
composition.
If a RNDIS is enabled with along with other debug functions
using the persist.vendor.usb.config.extra, set the
sys.usb.state to rndis(rndis,adb) irrespective of the config
that is being enabled to prevent the fallback.

Change-Id: Ib4e120895e62fc14ac11941b6b46cc7f0460ede8